refactor(export): lift the omission predicates out of export() (#2475)
The 259-line block between collectModifications and the output passes --
hasAnyUnreadableSourceRef, the deltaOnly empty short-circuit,
isOmittedFromOutput and mayNameOmittedRefs -- moves to
step-omission-predicates.ts. 698 -> 453.

The block moved verbatim except where becoming a function forced a change:

1. The short-circuit used to `return` out of export(). A free function
   cannot, so evaluateOmissionPredicates returns a tagged union and the
   caller returns `omission.result`. Tagged rather than an optional field,
   so reading isOmittedFromOutput off a short-circuited result is a compile
   error instead of an undefined at runtime.

   That rewrite is safe by measurement rather than by argument. c109f4126
   established the branch is not observable: disabling it leaves the suite
   green AND the emitted bytes and the whole stats object identical, because
   for a model with no edits the ordinary path arrives at the same place.
   Rewriting control flow nobody can observe cannot change behaviour.

2. `this.mutationView` arrives as a parameter -- it was the only `this` in
   the block.

Position is load-bearing, and the new module's header says so rather than
leaving it to be rediscovered: everything here reads what collectModifications
produced, and everything the output passes do with omitted refs reads what
this returns. Called earlier it answers from an unpopulated pass.

Three comments corrected, all made false by this move or already false:

- step-exporter.ts said the predicates "stay here instead of moving with
  it". They no longer stay here.
- step-collection.ts said they "stay in step-exporter.ts". Its actual
  argument was that putting them in step-collection would relocate a
  dependency rather than remove one -- true, and unaffected by a move to a
  third module. Reworded to say that, and to name where they went.
- A doc comment inside the moved block cited `if (options.visibleOnly &&
  this.dataStore.source)`. That code is `ctx.dataStore.source` at
  step-collection.ts:69 and has been in another file since the collection
  split; `this.` would have been doubly wrong in a module that has no
  `this`. Now cites the real spelling and location.

888/888 unchanged. tsc --noUnusedLocals clean. check-module-size exit 0;
step-omission-predicates.ts is 327 lines so it needs no allowlist row.

Note the shape of what is left: 453 lines, and the block just moved was 86%
comment (36 code lines of 259). The remaining candidates are
isGeometryEntity (~39) and the context builders, which lands near 400
rather than comfortably under it. Two of those builders share
`allocateExpressId: () => this.nextExpressId++`, and the counter's IDENTITY
has to survive any move -- capturing its value instead of the closure gives
each builder its own counter and silently double-allocates express ids.
